# internal/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hypervisor orchestration API for managing staged releases, deployments, and hyperuser access tooling.",
        "title": "OpenHack Hypervisor API",
        "contact": {},
        "version": "25.10.17.4"
    },
    "basePath": "/hypervisor",
    "paths": {
...
        "/hypervisor/hyperusers/login": {
            "post": {
                "description": "Validates hyperuser credentials and issues a 24-hour bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hyperusers Auth"
                ],
                "summary": "Hyperuser login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hyperusers.HyperUserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/hyperusers.HyperUserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errmsg._HyperUserInvalidPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errmsg._HyperUserWrongPassword"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errmsg._HyperUserNotExists"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errmsg._HyperUserInvalidPayload": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "username and password must be provided"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "errmsg._HyperUserNotExists": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "hyperuser does not exist"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "errmsg._HyperUserWrongPassword": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "username or password is incorrect"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 401
                }
            }
        },
        "errmsg._InternalServerError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "internal server error: \u003cdetails\u003e"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "hyperusers.HyperUserLoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "hyperusers.HyperUserLoginResponse": {
            "type": "object",
            "properties": {
                "hyperuser": {
                    "$ref": "#/definitions/models.HyperUser"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.HyperUser": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "HyperUserAuth": {
            "description": "Provide the hyperuser bearer token as `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}